+++ src/test/java/com/intrinio/api/StockExchangeApiTest.java
@@ -8,7 +8,6 @@ import com.intrinio.models.ApiResponseStockExchangeSecurities;
 import com.intrinio.models.ApiResponseStockExchangeStockPriceAdjustments;
 import com.intrinio.models.ApiResponseStockExchangeStockPrices;
 import com.intrinio.models.ApiResponseStockExchanges;
-import java.math.BigDecimal;
 import org.threeten.bp.LocalDate;
 import com.intrinio.models.StockExchange;
 import org.junit.Test;
@@ -41,7 +40,7 @@ public class StockExchangeApiTest {
         String city = null;
         String country = null;
         String countryCode = null;
-        BigDecimal pageSize = null;
+        Integer pageSize = null;
         ApiResponseStockExchanges response = api.getAllStockExchanges(city, country, countryCode, pageSize);
 
         // TODO: test validations
@@ -75,7 +74,7 @@ public class StockExchangeApiTest {
     public void getStockExchangePriceAdjustmentsTest() throws ApiException {
         String identifier = null;
         LocalDate date = null;
-        BigDecimal pageSize = null;
+        Integer pageSize = null;
         String nextPage = null;
         ApiResponseStockExchangeStockPriceAdjustments response = api.getStockExchangePriceAdjustments(identifier, date, pageSize, nextPage);
 
@@ -94,7 +93,7 @@ public class StockExchangeApiTest {
     public void getStockExchangePricesTest() throws ApiException {
         String identifier = null;
         LocalDate date = null;
-        BigDecimal pageSize = null;
+        Integer pageSize = null;
         String nextPage = null;
         ApiResponseStockExchangeStockPrices response = api.getStockExchangePrices(identifier, date, pageSize, nextPage);
 
@@ -113,7 +112,7 @@ public class StockExchangeApiTest {
     public void getStockExchangeRealtimePricesTest() throws ApiException {
         String identifier = null;
         String source = null;
-        BigDecimal pageSize = null;
+        Integer pageSize = null;
         String nextPage = null;
         ApiResponseStockExchangeRealtimeStockPrices response = api.getStockExchangeRealtimePrices(identifier, source, pageSize, nextPage);
 
@@ -131,7 +130,7 @@ public class StockExchangeApiTest {
     @Test
     public void getStockExchangeSecuritiesTest() throws ApiException {
         String identifier = null;
-        BigDecimal pageSize = null;
+        Integer pageSize = null;
         String nextPage = null;
         ApiResponseStockExchangeSecurities response = api.getStockExchangeSecurities(identifier, pageSize, nextPage);
